JBIDE-25386: Consolidate tests from 'org.hibernate.eclipse.console.test' into new test plugin 'org.jboss.tools.hibernate.orm.test' - Move test package 'mapping.onetomany' to be tested in org.jboss.tools.hibernate.orm.test.CoreMappingTest
