+++ tests/org.jboss.tools.hibernate.orm.test/res/project/src/core/onetomany/Parent.hbm.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <!DOCTYPE hibernate-mapping PUBLIC 
 	"-//Hibernate/Hibernate Mapping DTD 3.0//EN"
-	"http://hibernate.sourceforge.net/hibernate-mapping-3.0.dtd">
+	"http://www.hibernate.org/dtd/hibernate-mapping-3.0.dtd">
 
 <!-- 
 
@@ -13,7 +13,7 @@
 -->
 
 <hibernate-mapping 
-	package="mapping.onetomany">
+	package="core.onetomany">
 
 	<class name="Parent">
 		<id name="id"
